+++ packages/data-types/src/lib/interfaces/pet-shop-adoptions.interface.ts
@@ -1,7 +1,6 @@
 export interface PetShopAdoptions{
     npcData: {
         npcId: string;
-        nickname: string;
     }
     price: number;
     description: string;